Various fixes and refactoring

See merge request voyager/accelerator!16

+++ test/common/operations/Pooling.h
@@ -62,7 +62,8 @@ T *adaptive_avg_pool2d(std::map<std::string, std::any> &kwargs,
   const auto input_shape = get_shape(input);
 
   const auto output_size = op.kwargs().at("output_size").int_list().values();
-  std::vector<int> output_shape{output_size[0], output_size[1]};
+  std::vector<int> output_shape{static_cast<int>(output_size[0]),
+                                static_cast<int>(output_size[1])};
 
   int input_height = input_shape[2];
   int output_height = output_shape[0];
